Extract Utf8 marshalling to new InteropHelpers class
