+++ MonoGame.Framework/SDL/SDL2.cs
@@ -52,25 +52,6 @@ internal static class Sdl
     public static int Minor;
     public static int Patch;
 
-    public static unsafe string GetString(IntPtr handle)
-    {
-        if (handle == IntPtr.Zero)
-            return string.Empty;
-
-        var ptr = (byte*) handle;
-        while (*ptr != 0)
-            ptr++;
-
-        var len = ptr - (byte*) handle;
-        if (len == 0)
-            return string.Empty;
-
-        var bytes = new byte[len];
-        Marshal.Copy(handle, bytes, 0, bytes.Length);
-
-        return Encoding.UTF8.GetString(bytes);
-    }
-
     [Flags]
     public enum InitFlags
     {
@@ -234,7 +215,7 @@ internal static class Sdl
 
     public static string GetError()
     {
-        return GetString(SDL_GetError());
+        return InteropHelpers.Utf8ToString(SDL_GetError());
     }
 
     public static int GetError(int value)
@@ -263,7 +244,7 @@ internal static class Sdl
 
     public static string GetHint(string name)
     {
-        return GetString(SDL_GetHint(name));
+        return InteropHelpers.Utf8ToString(SDL_GetHint(name));
     }
 
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
@@ -510,7 +491,7 @@ internal static class Sdl
 
         public static string GetDisplayName(int index)
         {
-            return GetString(GetError(SDL_GetDisplayName(index)));
+            return InteropHelpers.Utf8ToString(GetError(SDL_GetDisplayName(index)));
         }
 
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
@@ -984,7 +965,7 @@ internal static class Sdl
 
         public static string GetMapping(IntPtr gamecontroller)
         {
-            return GetString(SDL_GameControllerMapping(gamecontroller));
+            return InteropHelpers.Utf8ToString(SDL_GameControllerMapping(gamecontroller));
         }
 
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
@@ -1002,7 +983,7 @@ internal static class Sdl
 
         public static string GetName(IntPtr gamecontroller)
         {
-            return GetString(SDL_GameControllerName(gamecontroller));
+            return InteropHelpers.Utf8ToString(SDL_GameControllerName(gamecontroller));
         }
     }
